Update BookController.java
Added token headers to endpoints

+++ src/main/java/wt/bookstore/backend/controllers/BookController.java
@@ -6,6 +6,7 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.web.bind.annotation.*;
 import wt.bookstore.backend.domains.Book;
+import wt.bookstore.backend.domains.User;
 import wt.bookstore.backend.dto.BookDto;
 import wt.bookstore.backend.dto.ChangeBookDto;
 import wt.bookstore.backend.dto.SaveBookDto;
@@ -13,6 +14,7 @@ import wt.bookstore.backend.mapping.BookDtoMapper;
 import wt.bookstore.backend.repository.IBookRepository;
 import wt.bookstore.backend.repository.ICopyRepository;
 import wt.bookstore.backend.repository.IReservationRepository;
+import wt.bookstore.backend.repository.IUserRepository;
 
 
 import java.util.List;
@@ -35,6 +37,9 @@ public class BookController {
     
     @Autowired
     private IReservationRepository reservationRepository;
+
+    @Autowired
+    private IUserRepository userRepository;
 
     @Autowired
     private BookDtoMapper bookMapper;
@@ -75,30 +80,67 @@ public class BookController {
      * @param saveBookDto ({@link wt.bookstore.backend.dto.SaveBookDto}) is generated from the json body in the POST request and contains the information needed to create a {@link wt.bookstore.backend.domains.Book} object.
      */
     @PostMapping("book/create")
-    public void create(@RequestBody SaveBookDto saveBookDto) {
+    public boolean create(@RequestBody SaveBookDto saveBookDto,
+                       @RequestHeader("Authentication") String token
+
+    ) {
+        Optional<User> userOptional = userRepository.findByToken(token);
+        if (userOptional.isEmpty()) {
+            return false;
+        }
+        User user = userOptional.get();
+        if (!user.isAdmin()){
+            return false;
+        }
         Book book = bookMapper.dtoToBook(saveBookDto);
         bookRepository.save(book);
+        return true;
     }
 
     @PutMapping("book/update/{id}")
-    public void update(@PathVariable long id, @RequestBody ChangeBookDto changeBookDto) {
+    public boolean update(@PathVariable long id, @RequestBody ChangeBookDto changeBookDto,
+                       @RequestHeader("Authentication") String token
+
+    ) {
+        Optional<User> userOptional = userRepository.findByToken(token);
+        if (userOptional.isEmpty()) {
+            return false;
+        }
+        User user = userOptional.get();
+        if (!user.isAdmin()){
+            return false;
+        }
         Optional<Book> optionalBook = bookRepository.findById(id);
-        if (optionalBook.isEmpty())
-            return;
+        if (optionalBook.isEmpty()) {
+            return false;
+        }
 
         optionalBook.get().setIsbn(changeBookDto.getIsbn());
         optionalBook.get().setTitle(changeBookDto.getTitle());
         optionalBook.get().setAuthor(changeBookDto.getAuthor());
 
         bookRepository.save(optionalBook.get());
+        return true;
     }
 
     /*
      * DELETE endpoints from here
      */
     @DeleteMapping("book/delete/{id}")
-    public void delete(@PathVariable long id) {
+    public boolean delete(@PathVariable long id,
+                       @RequestHeader("Authentication") String token
+
+    ) {
+        Optional<User> userOptional = userRepository.findByToken(token);
+        if (userOptional.isEmpty()) {
+            return false;
+        }
+        User user = userOptional.get();
+        if (!user.isAdmin()){
+            return false;
+        }
         bookRepository.deleteById(id);
+        return true;
     }
